+++ app/Http/Controllers/TypeCertificateController.php
@@ -78,7 +78,27 @@ class TypeCertificateController extends Controller
      */
     public function destroy(string $id)
     {
-        $type = TypeCertificate::find($id);
-        $type->delete();
+        try {
+            // Buscar el tipo de certificado
+            $type = TypeCertificate::find($id);
+            // Verificar si el tipo de certificado existe
+            if (!$type) {
+                $error = 404;
+                $message = "El tipo de certificado no se encontro";
+                return view('errors.encontro', compact('error', 'message'));
+            }
+            // Intentar eliminar el tipo de certificado
+            $type->delete();
+            // Si la eliminación es exitosa
+            $message = "tipo de certificado eliminado correctamente";
+            return view('errors.exitosa', compact('message'));
+        } catch (\Illuminate\Database\QueryException $e) {
+            // Si hay una violación de la restricción de clave foránea, enviar mensaje de error
+            if ($e->getCode() == 23000) {
+                $error = 400;
+            $message = "Error al intentar eliminar el tipo de certificado, ya que esta relacionada con otros registros";
+            return view('errors.middleware', compact('error', 'message'));
+            }
+        }
     }
 }
